frontend/requirements.txt
Brotli==1.0.9
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==3.0.1
chart-studio==1.1.0
click==8.1.3
colorama==0.4.6
dash==2.8.1
dash-auth==1.4.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Compress==1.13
Flask-SeaSurf==1.1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
plotly==5.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
retrying==1.3.4
six==1.16.0
tenacity==8.2.1
ua-parser==0.16.1
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.13.0
